+++ domain/root-post/factory/root-post-factory.ts
